requirements.txt
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
coloredlogs==15.0.1
datasets==2.11.0
diffusers==0.14.0
dill==0.3.6
evaluate==0.4.0
fastapi==0.95.0
filelock==3.9.0
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.4.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.2.0
Jinja2==3.1.2
lit==15.0.7
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==2.6.3
numpy==1.24.1
onnx==1.13.1
onnxruntime==1.14.1
opencv-contrib-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.3.0
protobuf==3.20.2
psutil==5.9.4
pyarrow==11.0.0
pydantic==1.10.7
pyngrok==5.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-triton==2.1.0+e650d3708b
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
responses==0.18.0
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tokenizers==0.13.3
torch==2.1.0.dev20230327+cu117
torchvision==0.16.0.dev20230327+cu117
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
